RUBY-1114 rename org.bson package to org.bson_ruby
This is to avoid naming conflicts with the Java driver when loaded concurrently
